Support large number of calls

+++ src/common/common_types.h
@@ -25,7 +25,7 @@ typedef struct avSRCountNode
     int size;
     uint64_t count; // How many time we detected the pattern; also size of list_calls
     uint64_t max_calls;
-    int *list_calls; // Which calls produced the pattern
+    uint64_t *list_calls; // Which calls produced the pattern
     int comm;
     int sendtype_size;
     int recvtype_size;
@@ -50,15 +50,15 @@ typedef struct avPattern
     // <n_ranks> ranks send to or receive from <n_peers> other ranks
     int n_ranks;
     int n_peers;
-    int n_calls;   // How many alltoallv calls have that pattern
+    uint64_t n_calls;   // How many collective calls have that pattern
     int comm_size; // Size of the communicator for which the pattern was detected. Not always used.
     struct avPattern *next;
 } avPattern_t;
 
 typedef struct avCallPattern
 {
-    int n_calls;
-    int *calls;
+    uint64_t n_calls;
+    uint64_t *calls;
     avPattern_t *spatterns;
     avPattern_t *rpatterns;
     struct avCallPattern *next;
@@ -66,8 +66,8 @@ typedef struct avCallPattern
 
 typedef struct caller_info
 {
-    int n_calls;
-    int *calls;
+    uint64_t n_calls;
+    uint64_t *calls;
     char *caller;
     struct caller_info *next;
 } caller_info_t;
